==> Quinte/DataFormattingScript/1_format_edm_vitals_script.ipynb <==
#this file formats the edm raw data file 

import pandas as pd
import numpy as np

df1 = pd.read_excel(
     '../ectas_edm_vitals.xlsx',
     engine='openpyxl', sheet_name = 'edm_vitals'
)

df1

del df1['screen']


rows = len(df1.index)
df1

df2 = pd.DataFrame(data=[], columns= ['pt_identifier', 'Date', 'Time', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat'])

df2

def count_consecutive(df1, i, last_idx):
    count = 1
    
    while np.sum(df1.iloc[i][0:3] == df1.iloc[i + 1][0:3]) == 3:
        count += 1
        i += 1
        if i == last_idx:
            break
    
    return count

def find_index(key_list, key):
    idx = -1
    for x in range(len(key_list)):
        if key_list[x] == key:
            idx = x
        
    return idx

def extract_vitals(df1, i, consecutive):
    p_id = df1.iloc[i][0]
    date = df1.iloc[i][1]
    time = df1.iloc[i][2]
       
    key_list  = list(df1.iloc[i:i+consecutive,3])
    value_list = list(df1.iloc[i:i+consecutive,4])
    
    bp_idx = find_index(key_list, 'N.000300.1')
    pulse_idx = find_index(key_list, 'N.000280.1')
    temp_idx = find_index(key_list, 'N.000260.1')
    rr_idx = find_index(key_list, 'N.000290.1')
    o2sat_idx = find_index(key_list, 'N.000320.1')

    
    bp_combined = value_list[bp_idx] if bp_idx != -1 else ''
    
    bp1, bp2 = bp_combined.split('/') if bp_combined != '' else ['','']
    
    pulse = value_list[pulse_idx] if pulse_idx != -1 else ''
    temp = value_list[temp_idx] if temp_idx != -1 else ''
    rr = value_list[rr_idx] if rr_idx != -1 else ''
    o2sat = value_list[o2sat_idx] if o2sat_idx != -1 else ''

    
    p_vitals = [p_id, date, time, bp1, bp2, pulse, temp, rr, o2sat]
    return p_vitals

i = 0
while i < rows:
    consecutive = count_consecutive(df1, i, rows - 1)
    p_vitals = extract_vitals(df1, i, consecutive)
    df2 = df2.append({'pt_identifier': p_vitals[0], 'Date': p_vitals[1],
                      'Time': p_vitals[2], 'BP1': p_vitals[3], 'BP2': p_vitals[4],
                      'Pulse': p_vitals[5], 'Temp': p_vitals[6], 'RR': p_vitals[7],
                      'O2Sat': p_vitals[8]}, ignore_index=True)
    i += consecutive
  
    


df2.iloc[17660:]

df2.to_csv('1.csv') #formatted edm vitals


# run time approx 5 minutes for 75,000 rows
