==> Quinte/DataFormattingScript/2_format_nur_vitals.ipynb <==
#this file formats the raw NUR vital file

import pandas as pd
import numpy as np
import math

df1 = pd.read_excel(
     '../nur_vitals.xlsx',
     engine='openpyxl', sheet_name = 'NUR_vitals'
)

df1

del df1['td.locat']
del df1["Temp_Source"]
del df1["Interv"]

rows = len(df1.index)
df1

df2 = pd.DataFrame(data=[], columns= ['pt_identifier', 'Date', 'Time', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat'])

df2

rows

def extract_vitals(df1, i):
    p_id = df1.iloc[i][0]
    date = df1.iloc[i][1]
    time = df1.iloc[i][2]
    pulse = df1.iloc[i][4]
    rr = df1.iloc[i][5]
    o2sat = df1.iloc[i][6]
    temp = df1.iloc[i][7]
    
    if df1.iloc[i][3]:
        pre_bp = df1.iloc[i][3]
        if type(pre_bp) == float and math.isnan(pre_bp)==False:
            manual_bp = str(pre_bp)
            x = manual_bp.split("/")
            if len(x)==2:
                bp1 = float(x[0])
                bp2 = float(x[1])
            else:
                bp1 = None
                bp2 = None
        elif type(pre_bp) == str:
            manual_bp = str(pre_bp)
            if "E" in manual_bp:
                manual_bp = manual_bp.replace("E", "")
                manual_bp = manual_bp.strip()
            x = manual_bp.split("/")
            if len(x)==2:
                bp1 = float(x[0])
                bp2 = float(x[1])
            else:
                bp1 = None
                bp2 = None
        else:
            bp1 = None
            bp2 = None

    elif df1.iloc[i][8] and df1.iloc[i][9] and math.isnan(df1.iloc[i][8])==False and math.isnan(df1.iloc[i][9])==False:
        bp1 = df1.iloc[i][8]
        bp2 = df1.iloc[i][9]

    else:
        bp1 = None
        bp2 = None
    


    
    p_vitals = [p_id, date, time, bp1, bp2, pulse, temp, rr, o2sat]
    return p_vitals

i = 0
while i < rows:
    p_vitals = extract_vitals(df1, i)
    df2 = df2.append({'pt_identifier': p_vitals[0], 'Date': p_vitals[1],
                      'Time': p_vitals[2], 'BP1': p_vitals[3], 'BP2': p_vitals[4],
                      'Pulse': p_vitals[5], 'Temp': p_vitals[6], 'RR': p_vitals[7],
                      'O2Sat': p_vitals[8]}, ignore_index=True)
    i += 1
  
    


df2.iloc[105500:105550]

df2.to_csv('2.csv')#formatted_good_nur_vitalsall.csv


# run time approx 5 minutes for 75,000 rows
