==> Quinte/DataFormattingScript/3_combine_all_files.ipynb <==
#this file combines the EDM and NUR vitals, as well as patient data

import pandas as pd


df1 = pd.read_csv(
     '../FormattedData/1.csv',
)

df1

del df1["Unnamed: 0"]

df1

df2 = pd.read_csv(
     '../FormattedData/2.csv',
)

df2

del df2["Unnamed: 0"]

df2

df3 = pd.read_excel(
     '../ectas_edm_vitals.xlsx',
     engine='openpyxl', sheet_name = 'ectas_vitals'
)

df3

del df3["Unnamed: 9"]

df3

df3 = df3.rename(columns={"Triage D": "Date"})


df3 = df3.rename(columns={"tTm": "Time"})


df3

df3 = df3.drop(df3.index[8127:])

df3

frames = [df1, df2, df3]

df4 = pd.concat(frames, ignore_index=True)

df4

df4.to_csv('3.csv')#master_without_labs

df_master = pd.read_excel(
     '../MF_lact_WC_Diag.xlsx',
     engine='openpyxl', sheet_name = 'pt_master_file'
)

df_master

header_list = ["pt_identifier", "Date", "Time", "BP1", "BP2", "Pulse", "Temp", "RR", "O2Sat", "Admission_Date", "Discharge_Date", "Age_Code", "Age_Number", "Gender", "Weight", "LDate", "LTime", "LResult", "LRange", "WDate", "WTime", "WResult", "Diag_Code", "Diag_desc", "Diag_prefix", "Diag_type"]

df4 = df4.reindex(columns = header_list)          

df4

big_pt = df4["pt_identifier"].tolist()
small_pt = df_master["de_identified"].tolist()

for i in range(len(big_pt)):
    for j in range(len(small_pt)):
        if big_pt[i] == small_pt[j]:
            df4["Admission_Date"][i] = df_master.iloc[j][1]
            df4["Discharge_Date"][i] = df_master.iloc[j][2]
            df4["Age_Code"][i] = df_master.iloc[j][3]
            df4["Age_Number"][i]= df_master.iloc[j][5]
            df4["Gender"][i] = df_master.iloc[j][6]
            df4["Weight"][i] = df_master.iloc[j][7]
            break

df4

df_diag = pd.read_excel(
     '../MF_lact_WC_Diag.xlsx',
     engine='openpyxl', sheet_name = 'diagnoses'
)

df_diag

diag_id = df_diag["de_identified"].tolist()

for i in range(len(big_pt)):
    for j in range(len(diag_id)):
        if big_pt[i] == diag_id[j]:
            df4["Diag_Code"][i] = df_diag.iloc[j][3]
            df4["Diag_desc"][i] = df_diag.iloc[j][4]
            df4["Diag_prefix"][i] = df_diag.iloc[j][5]
            df4["Diag_type"][i]= df_diag.iloc[j][7]
            break


header_list = ["pt_identifier", "Date", "Time", "BP1", "BP2", "Pulse", "Temp", "RR", "O2Sat", "Admission_Date", "Discharge_Date", "Age_Code", "Age_Number", "Gender", "Weight", "Diag_Code", "Diag_desc", "Diag_prefix", "Diag_type"]

df4 = df4.reindex(columns = header_list)

df4.to_csv('4.csv')#all_vitals_info_diag


df4
