==> Quinte/DataFormattingScript/4_add_in_lactate.ipynb <==
#this file adds in lactate values

import pandas as pd
import numpy as np

df1 = pd.read_csv(
     '../FormattedData/4.csv',
)

df1

df2 = pd.read_excel(
     '../MF_lact_WC_Diag.xlsx',
     engine='openpyxl', sheet_name = 'lactate'
)

df2

del df2["status"]
del df2["test_desc"]
del df2["test"]
del df2["Unnamed: 10"]

df2

df2 = df2.rename(columns = {"date": "Date", "time": "Time"})

df2 = df2.rename(columns = {"result":"Lactate_Value", "L_H_VL_VH":"Lactate_High_Low", "ref_range": "Lactate_Ref_Range"})

del df2["ref_range2"]

df2

df2 = df2.drop(df2.index[16384:])

df2 = df2.rename(columns = {"result":"Lactate_Value", "L_H_VL_VH":"Lactate_High_Low", "ref_range": "Lactate_Ref_Range"})

df2

df1["Lactate_Value"] = np.nan
df1["Lactate_High_Low"] = np.nan
df1["Lactate_Ref_Range"] = np.nan


df1_df2 = df1.append(df2, sort=False, ignore_index=True)

df1_df2

del df1_df2["Unnamed: 0"]

df1_df2

df1_df2.to_csv('5.csv')#vitals_info_lactate
