==> Quinte/DataFormattingScript/5_add_in_WBC.ipynb <==
#this file adds in wbc values


import pandas as pd
import numpy as np

df1 = pd.read_csv(
     '../FormattedData/5.csv',
)

df1

df2 = pd.read_excel(
     '../MF_lact_WC_Diag.xlsx',
     engine='openpyxl', sheet_name = 'WC'
)

df2

del df2["status"]
del df2["test_desc"]
del df2["test"]
del df2["ref_range2"]

df2

df2 = df2.rename(columns = {"date": "Date", "time": "Time"})

df2 = df2.rename(columns = {"result":"WBC_Value", "L_H_VL_VH":"WBC_High_Low", "ref_range": "WBC_Ref_Range"})

df2

df1["WBC_Value"] = np.nan
df1["WBC_High_Low"] = np.nan
df1["WBC_Ref_Range"] = np.nan


df1

del df1["Unnamed: 0"]

df1

df1_df2 = df1.append(df2, sort=False, ignore_index=True)

df1_df2

df1_df2.to_csv('6.csv')#vitals_info_lactate_WBC
