==> Quinte/DataFormattingScript/6_sort_by_pt_identifier.ipynb <==
#this file sorts the dataset by patient identifier

import pandas as pd

df1 = pd.read_csv(
     '../FormattedData/6.csv',
)

df1

del df1["Unnamed: 0"]

df1

df1.sort_values(by=["pt_identifier"], inplace = True)

df1

df1.reset_index(inplace = True, drop = True)

df1

df1.to_csv('7.csv')#vitals_info_lactate_WBC_sorted_pt
