==> Quinte/DataFormattingScript/7_sort_by_time_normalize.ipynb <==
#this file normalizes the 'date' into relative minutes

import pandas as pd

df1 = pd.read_csv(
     '../FormattedData/7.csv',
)

df1

del df1["Unnamed: 0"]

df1

del df1["Admission_Date"]
del df1["Discharge_Date"]
df1

type(df1.iloc[0][1])

num_rows = 0
counter = 0
while counter<77:
    if df1.iloc[num_rows][2] == "UNK":
        df1.drop(df1.index[num_rows], inplace = True)
        #print(num_rows)
        counter+=1
    num_rows+=1

df1

df1.reset_index(inplace = True, drop = True)

df1

df1.iloc[2698][2]

rows = len(df1.index)
rows

import datetime

df1.iloc[312757][1]

#get date and time into lists per patient;
#do NOT change order
#find lowest date in that list
#make another list with same indexing of relative time (lowest is 0)(units in minutes --> 0, 10, 20, 70, 100)
#combine all these lists (append to main)
#put it in df unsorted
#sort df
def date_to_number(df1, i, consecutive):
    current_row = i
    date_list = []
    time_list = []

    for x in range(consecutive):
        date_list.append(df1.iloc[current_row][1])
        time_list.append(df1.iloc[current_row][2])
        current_row+=1
    
    date_num_list = []

    for y in range(len(date_list)):
        #print(date_list[y])
        x = (str(date_list[y])).split("/")
        date_int = [int(x[0]), int(x[1]), int(x[2])]
        date_num_list.append(date_int)

    time_num_list = []
    for z in range(len(time_list)):
        #print(time_list[z])
        a = float(time_list[z])
        b = int(a)

        if b<100:
            time_num_list.append([0,b])

        elif b<=1000 and b>100:
            c = b%100
            d = int(str(b)[0])
            time_num_list.append([d,c])
        
        else:
            e = b%100
            f = int(str(b)[0:2])
            time_num_list.append([f,e])
    
    # print(len(date_num_list))
    # print(len(time_num_list))
    date_time_list = []
    for xy in range(consecutive):
        new_date = datetime.datetime(date_num_list[xy][2], date_num_list[xy][1], date_num_list[xy][0], time_num_list[xy][0], time_num_list[xy][1])
        date_time_list.append(new_date)
        
    #print(datetime.datetime(2020, 5, 17))
    min = date_time_list[0]
    low_index = 0

    for yz in range(1, len(date_time_list)):
        if date_time_list[yz]<min:
            min = date_time_list[yz]
            low_index = yz
    
    normalized_time = []

    for xyz in range(len(date_time_list)):
        date1 = date_time_list[xyz]
        date0 = date_time_list[low_index]
        result_days = int( (date1-date0).days)
        result_seconds = int( (date1-date0).seconds)
        good_result = float( (((86400*result_days)+result_seconds)/60) )
        normalized_time.append(good_result)

    return normalized_time



def count_consecutive(df1, i, last_idx):
    count = 1

    while i<rows-1 and df1.iloc[i][0] == df1.iloc[i+1][0] :
        count += 1
        i += 1
        print(i)
        if i == last_idx:
            print("hiiiiiiiii")
            break
    
    return count    

rows

final_array = []
i=0
while i<rows:
    consecutive = count_consecutive(df1, i, rows-1)
    #print(i)
    each_array = date_to_number(df1, i, consecutive)
    for t in each_array:
        final_array.append(t)
    i+=consecutive
    


len(final_array)
df1["new_date"] = final_array
df1

#df1.sort_values(by=["new_date"], inplace = True)

df1.reset_index(inplace = True, drop = True)

df1.columns

df1

del df1["Date"]
del df1["Time"]


df1 = df1[['pt_identifier', 'new_date', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat',
       'Age_Code', 'Age_Number', 'Gender', 'Weight', 'Diag_Code', 'Diag_desc',
       'Diag_prefix', 'Diag_type', 'Lactate_Value', 'Lactate_High_Low',
       'Lactate_Ref_Range', 'WBC_Value', 'WBC_High_Low', 'WBC_Ref_Range']]

df1

df1.to_csv('8.csv')#vitals_info_lactate_WBC_sorted_pt_and_time, normalized too
