==> Quinte/DataFormattingScript/10_diag_to_numerical.ipynb <==
#this file turns descriptions of sepsis into numerical values (0 and 1)
#also one hot encodes the diag_type from m, 1, 2, 3, to 4 columns with 0's and 1's with proper columns

import pandas as pd
import math

df1 = pd.read_csv(
     '../FormattedData/10.csv',
)

df1

del df1["Unnamed: 0"]

df1

rows = len(df1)
rows

print (df1.iloc[5][11])

i=0
while i<rows:
    if df1.iloc[i][11] == "Non sepsis patient":
        df1.at[i, "Diag_desc"] = 0
    else:
        df1.at[i, "Diag_desc"] = 1
    i+=1
    


df1

#df1.to_csv('stage5_1_diag_to_numerical')


del df1["Diag_prefix"]

df1

import math

print (df1.iloc[1763][12])
print( type( df1.iloc[1763][12]))
print (math.isnan(df1.iloc[1][12]))

i=0
most = []
pre = []
post = []
secondary = []
while i<rows:
   # print(df1.iloc[i][12])
    if df1.iloc[i][12] == "M":
            most.append(1)
            pre.append(0)
            post.append(0)
            secondary.append(0)
    elif df1.iloc[i][12] == "1":
            most.append(0)
            pre.append(1)
            post.append(0)
            secondary.append(0)
    elif df1.iloc[i][12] == "2":
            most.append(0)
            pre.append(0)
            post.append(1)
            secondary.append(0)
    elif df1.iloc[i][12] == "3":
            most.append(0)
            pre.append(0)
            post.append(0)
            secondary.append(1)
    elif math.isnan(df1.iloc[i][12]):
            most.append(0)
            pre.append(0)
            post.append(0)
            secondary.append(0)
    i+=1
    

del df1["Diag_type"]
df1

df1["Most_responsible"] = most
df1["pre_admit"] = pre
df1["post_admit"] = post
df1["secondary"] = secondary

df1 = df1[['pt_identifier', 'new_date', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat',
       'Age_Number', 'Gender', 'Weight', 'Diag_desc',
       'Most_responsible', 'pre_admit', 'post_admit', 'secondary', 'Lactate_Value', 'Lactate_High_Low',
       'Lactate_Ref_Range', 'WBC_Value', 'WBC_High_Low', 'WBC_Ref_Range']]

df1

df1.to_csv('11.csv')
