==> Quinte/DataFormattingScript/11_adjusted_labs.ipynb <==
#this file turns the lab value columns into one adjusted lab value by dividing the real lab value by the high of the reference range value

import pandas as pd
import math

df1 = pd.read_csv(
     '../FormattedData/11.csv'
)

df1

del df1["Lactate_High_Low"]

del df1["WBC_High_Low"]

del df1["Unnamed: 0"]

df1

rows = len(df1)
rows

print (df1.iloc[15][18])
print (type(df1.iloc[15][18]))

print(df1.iloc[15][19])
print (type(df1.iloc[15][19]))


def get_higher_value(text):
    newer = text.split("-")[-1]
    return float(newer)


i=0
#16, 17, 18, 19
lactate_list = []
wbc_list = []
while i<rows:
    value1 = df1.iloc[i][16]
    try:
        value1 = float(value1)
    except:
        if "<" in value1:
            value1 = value1.split("<")[-1]
            value1 = value1.strip()
            value1 = float(value1)
        elif ">" in value1:
            value1 = value1.split(">")[-1]
            value1 = value1.strip()
            value1 = float(value1)

    if not math.isnan(value1):
        init_l = value1
        string_l = df1.iloc[i][17]
        newer_float_l = get_higher_value(string_l)

        final_l = init_l/newer_float_l
        lactate_list.append(final_l)
    else:
        lactate_list.append(float(0))
    

    value2 = df1.iloc[i][18]
    try:
        value2 = float(value2)
    except:
        if "<" in value2:
            value2 = value2.split("<")[-1]
            value2 = value2.strip()
            value2 = float(value2)
        elif ">" in value2:
            value2 = value2.split(">")[-1]
            value2 = value2.strip()
            value2 = float(value2)

    if not math.isnan(value2):
        init_w = value2
        string_w = df1.iloc[i][19]
        newer_float_w = get_higher_value(string_w)

        final_w = init_w/newer_float_w
        wbc_list.append(final_w)
    else:
        wbc_list.append(float(0))
    

    i+=1
    


df1["Lactate_adjusted"] = lactate_list
df1["WBC_adjusted"] = wbc_list

del df1["Lactate_Value"]
del df1["Lactate_Ref_Range"]
del df1["WBC_Value"]
del df1["WBC_Ref_Range"]

df1 = df1[['pt_identifier', 'new_date', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat',
       'Age_Number', 'Gender', 'Weight', 'Diag_desc',
       'Most_responsible', 'pre_admit', 'post_admit', 'secondary', 'Lactate_adjusted', 'WBC_adjusted']]

df1


df1.to_csv('12.csv')
