==> Quinte/DataFormattingScript/12_separate_old_data.ipynb <==
#this file looks at the time and separates them if the value is over 10,000 minutes away (1 week)

import pandas as pd

df1 = pd.read_csv(
     '../FormattedData/12.csv)'
)

df1

del df1["Unnamed: 0"]

df1

rows = len(df1)
rows

print (df1.iloc[0][1])


i=1
date_list = []
date_list.append(df1.iloc[0][1])
while i<rows:
    value1 = df1.iloc[i][1]
    value0 = df1.iloc[i-1][1] 

    if (value1-value0) > 10080:
        subtracted = df1.iloc[i][1]
        date_list.append(0.0)
        new_idx = i+1

        while df1.iloc[new_idx][1] != 0.0 and (df1.iloc[new_idx][1]-df1.iloc[new_idx-1][1])<10080:
            changed_v = df1.iloc[new_idx][1]
            date_list.append(changed_v - subtracted)
            new_idx += 1
            i += 1
    else:
        date_list.append(value1)
    

    i+=1
    


del df1["new_date"]
df1["new_date"] = date_list

df1 = df1[['pt_identifier', 'Diag_desc', 'new_date', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat',
       'Age_Number', 'Gender', 'Weight',
       'Most_responsible', 'pre_admit', 'post_admit', 'secondary', 'Lactate_adjusted', 'WBC_adjusted']]

df1


df1.head(50)

df1.to_csv('13.csv')
