==> Quinte/DataFormattingScript/13_one_hot_encode_diag_gender.ipynb <==
#this file one hot encodes sepsis identification and gender

import pandas as pd
import math

df1 = pd.read_csv(
     '../FormattedData/13.csv)'
)

df1

del df1["Unnamed: 0"]

df1

rows = len(df1)
rows

print (df1.iloc[64][10])
print (type(df1.iloc[64][10]))
#row column


i=0
positive = []
negative = []
male = []
female = []


while i<rows:
    diag = df1.iloc[i][1]
    if diag == 0:
        negative.append(1)
        positive.append(0)
    elif diag == 1:
        positive.append(1)
        negative.append(0)
    else:
        print("rows", i)
        print (diag)
        print(type(diag))
        raise ValueError
    
    gender = df1.iloc[i][10]
    if gender == "M" or gender == "m":
        male.append(1)
        female.append(0)
    elif gender == "F" or gender == "f":
        female.append(1)
        male.append(0)
    else:
        try:
            value1 = float(gender)
            if math.isnan(value1):
                male.append(0)
                female.append(0)
            else:
                print(value1)
                print("HERE")
                raise ValueError
        except:
            print("cant convert to float")
            print("rows", i)
            print(gender)
            print(type(gender))
            raise ValueError


   

    i+=1
    


del df1["Diag_desc"]
df1["positive_diag"] = positive
df1["negative_diag"] = negative

del df1["Gender"]
df1["Male"] = male
df1["Female"] = female

df1 = df1[['pt_identifier', 'positive_diag', 'negative_diag', 'new_date', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat',
       'Age_Number', 'Male', 'Female', 'Weight',
       'Most_responsible', 'pre_admit', 'post_admit', 'secondary', 'Lactate_adjusted', 'WBC_adjusted']]

df1


df1.to_csv('14.csv')
