==> Quinte/DataFormattingScript/15_del_weight_find_sepsis_num.ipynb <==
#this file deletes the weight category as it is not indicative or useful, (also, 20% of the rows do not have weight anyway)

import pandas as pd
#import modin.pandas as pd

df1 = pd.read_csv(
     '../FormattedData/15.csv'
)

df1

del df1["Unnamed: 0"]

df1.head(10)

rows = len(df1)
rows

del df1["Weight"]

df1

print (df1.iloc[65][1])
print ((df1.iloc[65][3]))
#row column


i=0
count = 0

while i<rows:
    value1 = df1.iloc[i][1]
    value2 = df1.iloc[i][3]

    if value1 == 1 and (value2 == 0.0 or value2 == 0):
        count+=1
    
    i+=1

print(count)
    

i=0
count2 = 0

while i<rows:
    value2 = df1.iloc[i][3]

    if (value2 == 0.0 or value2 == 0):
        count2+=1
    
    i+=1

print(count2)

df1

df1.to_csv('../FormattedData/16.csv') #took out weight
